requirements.txt
torch
transformers
diffusers
Pillow
torchvision